Refactor: add admin validation in delete users

+++ tests/app_test.py
@@ -72,7 +72,9 @@ def registerPassenger2():
         },
     )
     return response
-
+def adminToken():
+    token = token_handler.create_access_token(500, "admin")
+    return token;
 
 def addAdressClient(endpoint):
     response = client.patch(
@@ -126,8 +128,8 @@ def test_when_app_has_2_passengers_then_get_passengers_return_2_users():
     print(data)
     assert len(data) == 2
     assert data[0]["id"] == data1["id"]
-    client.delete("/users/" + str(data[0]["id"]), json={"user_type": "passenger"})
-    client.delete("/users/" + str(data[1]["id"]), json={"user_type": "passenger"})
+    client.delete("/users/" + str(data[0]["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
+    client.delete("/users/" + str(data[1]["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_app_has_2_user_then_get_users_return_2_users():
@@ -141,8 +143,8 @@ def test_when_app_has_2_user_then_get_users_return_2_users():
     data = response.json()
     assert len(data) == 2
     print(data)
-    client.delete("/users/" + str(data[0]["id"]), json={"user_type": "passenger"})
-    client.delete("/users/" + str(data[1]["id"]), json={"user_type": "passenger"})
+    client.delete("/users/" + str(data[0]["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
+    client.delete("/users/" + str(data[1]["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_getting_a_passenger_info_that_existis_then_it_returns_it():
@@ -153,7 +155,7 @@ def test_when_getting_a_passenger_info_that_existis_then_it_returns_it():
     assert response.status_code == status.HTTP_200_OK
     data1 = response.json()
     print("test ", data1)
-    client.delete("/users/" + str(data["id"]), json={"user_type": "passenger"})
+    client.delete("/users/" + str(data["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_creating_a_passenger_with_not_registered_email_creates_the_user():
@@ -168,7 +170,7 @@ def test_when_creating_a_passenger_with_not_registered_email_creates_the_user():
     assert data["email"] == "sol@gmail.com"
     assert "id" in data
     assert data["picture"] == "picture"
-    client.delete("/users/" + str(data["id"]), json={"user_type": "passenger"})
+    client.delete("/users/" + str(data["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_creating_passenger_with_registered_email_does_not_create_the_passenger():
@@ -179,7 +181,7 @@ def test_when_creating_passenger_with_registered_email_does_not_create_the_passe
     data = response.json()
     assert data["detail"] == "The passenger already exists"
 
-    client.delete("/users/1", json={"user_type": "passenger"})
+    client.delete("/users/1", json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_creating_a_driver_withnot_registered_email_creates_the_driver():
@@ -194,7 +196,7 @@ def test_when_creating_a_driver_withnot_registered_email_creates_the_driver():
     assert data["age"] == 22
     assert "id" in data
 
-    client.delete("/users/" + str(data["id"]), json={"user_type": "driver"})
+    client.delete("/users/" + str(data["id"]), json={"user_type": "driver"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_creating_driver_with_registered_email_doesnot_create_the_driver():
@@ -203,7 +205,7 @@ def test_when_creating_driver_with_registered_email_doesnot_create_the_driver():
     assert response.status_code == status.HTTP_409_CONFLICT, response.text
     data = response.json()
     assert data["detail"] == "The driver already exists"
-    client.delete("/users/1", json={"user_type": "driver"})
+    client.delete("/users/1", json={"user_type": "driver"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_a_passanger_exists_and_add_the_address_then_the_addres_is_add_it():
@@ -215,7 +217,7 @@ def test_when_a_passanger_exists_and_add_the_address_then_the_addres_is_add_it()
     assert response.status_code == status.HTTP_200_OK, response.text
     assert data[1]["default_address"] == "Av avellaneda 123"
 
-    client.delete("/users/" + str(id), json={"user_type": "passenger"})
+    client.delete("/users/" + str(id), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_Passenger_not_exist_and_adds_address_then_the_addres_isnot_addit():
@@ -235,7 +237,7 @@ def test_when_driver_exists_and_adds_carInfo_the_carInfo_is_addit():
     assert data[1]["license_plate"] == "ABC123"
     assert data[1]["car_model"] == "Ford K"
 
-    client.delete("/users/" + str(id), json={"user_type": "driver"})
+    client.delete("/users/" + str(id), json={"user_type": "driver"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_driver_not_exists_and_adds_carInfo_the_carInfo_isnot_addit():
@@ -274,7 +276,7 @@ def test_when_gettion_info_from_existing_user_then_returns_the_information():
     assert data2["phone_number"] == data["phone_number"]
     assert data2["age"] == data["age"]
 
-    client.delete("/users/" + str(data["id"]), json={"user_type": "passenger"})
+    client.delete("/users/" + str(data["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_login_to_register_user_with_validad_data_then_it_should_return_token():
@@ -290,7 +292,7 @@ def test_when_login_to_register_user_with_validad_data_then_it_should_return_tok
 
     assert actual["user_id"] == expected["user_id"]
     assert actual["user"] == expected["user"]
-    client.delete("/users/" + str(1), json={"user_type": "passenger"})
+    client.delete("/users/" + str(1), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_login_register_user_with_invalid_email_then_it_should_not_return_token():
@@ -317,7 +319,7 @@ def test_when_getting_profile_for_passenger_that_exists_it_should_do_it():
     assert data[0]["phone_number"] == "12345678"
     assert data[0]["email"] == "agus@gmail.com"
     assert data[1]["default_address"] is None
-    client.delete("/users/" + str(data1["id"]), json={"user_type": "passenger"})
+    client.delete("/users/" + str(data1["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_update_passenger_info_it_should_do_it():
@@ -339,7 +341,7 @@ def test_when_update_passenger_info_it_should_do_it():
 
     assert data[0]["age"] == 25
     assert data[1]["default_address"] == "example"
-    client.delete("/users/" + str(data1["id"]), json={"user_type": "passenger"})
+    client.delete("/users/" + str(data1["id"]), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_update_driver_info_it_should_do_it():
@@ -362,7 +364,7 @@ def test_when_update_driver_info_it_should_do_it():
     assert data[0]["phone_number"] == "436278"
     assert data[1]["model_car"] == "Audi"
 
-    client.delete("/users/" + str(data1["id"]), json={"user_type": "driver"})
+    client.delete("/users/" + str(data1["id"]), json={"user_type": "driver"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_update_driver_that_not_exist_it_should_not_do_it():
@@ -412,7 +414,7 @@ def test_when_login_with_google_with_user_not_register_then_it_returns_token():
 
     assert actual["user_id"] == expected["user_id"]
     assert actual["user"] == expected["user"]
-    client.delete("/users/" + str(1), json={"user_type": "passenger"})
+    client.delete("/users/" + str(1), json={"user_type": "passenger"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 # def test_when_login_with_google_when_user_register_normaly_does_not_return_token():
@@ -444,7 +446,7 @@ def test_when_scoring_a_driver_that_exist_it_does_it():
     score_expected = (3 + 4) / 2
 
     assert score_expected == data1["score"]
-    client.delete("/users/" + str(data["id"]), json={"user_type": "driver"})
+    client.delete("/users/" + str(data["id"]), json={"user_type": "driver"}, headers={"Authorization": f"Bearer {adminToken()}"})
 
 
 def test_when_scoring_a_driver_that_does_not_exist_it_does_not_do_it():
@@ -460,3 +462,12 @@ def test_when_scoring_a_driver_that_does_not_exist_it_does_not_do_it():
     assert response.status_code == status.HTTP_404_NOT_FOUND, response.text
     data = response.json()
     assert data["detail"] == "The driver does not exists"
+
+def test_when_deleting_a_user_with_no_autherization_it_does_bot_do_it():
+    token = token_handler.create_access_token(100, "user")
+    response = client.delete("/users/100", json={"user_type": "driver"}, headers={"Authorization": f"Bearer {token}"})
+    assert response.status_code == status.HTTP_401_UNAUTHORIZED
+
+    data = response.json()
+
+    assert data['detail'] == "The user is not authorize"
